requirements.txt
torch==1.11.0
torchvision==0.11.0
pydantic==1.9.0
tensorfn==0.1.23
opencv-python==4.6.0.66
lmdb==1.3.0
Shapely==1.8.1.post1
Pillow==9.0.1
black==23.3.0
numpy==1.21.4
rapidfuzz==2.9.0
Polygon3==3.0.9.1
editdistance==0.6.2
